requirements.txt
django==3.0.3
djangorestframework==3.11.0
gunicorn==19.9.0
whitenoise==4.1.2
dj-database-url==0.5.0